fix enums usage

+++ src/ChickenAPI/Features/Shops/ShopEventHandler.cs
@@ -63,16 +63,16 @@ namespace ChickenAPI.Game.Features.Shops
 
             switch (player.GetDignityIcon())
             {
-                case (int)CharacterDignity.BluffedNameOnly:
+                case CharacterDignity.BluffedNameOnly:
                     percent = 1.1;
                     typeshop = 110;
                     break;
-                case (int)CharacterDignity.NotQualifiedFor:
+                case CharacterDignity.NotQualifiedFor:
                     percent = 1.2;
                     typeshop = 120;
                     break;
-                case (int)CharacterDignity.Useless:
-                case (int)CharacterDignity.StupidMinded:
+                case CharacterDignity.Useless:
+                case CharacterDignity.StupidMinded:
                     percent = 1.5;
                     typeshop = 150;
                     break;
@@ -263,14 +263,14 @@ namespace ChickenAPI.Game.Features.Shops
             double percent = 1.0;
             switch (player.GetDignityIcon())
             {
-                case (int)CharacterDignity.BluffedNameOnly:
+                case CharacterDignity.BluffedNameOnly:
                     percent = 1.10;
                     break;
-                case (int)CharacterDignity.NotQualifiedFor:
+                case CharacterDignity.NotQualifiedFor:
                     percent = 1.20;
                     break;
-                case (int)CharacterDignity.Useless:
-                case (int)CharacterDignity.StupidMinded:
+                case CharacterDignity.Useless:
+                case CharacterDignity.StupidMinded:
                     percent = 1.5;
                     break;
             }
